--- packages/app/src/redirect.ts
@@ -1,14 +1,0 @@
-export const redirectRoutes = [
-  {
-    path: '/Design-System/About-the-Design-System',
-    to: '/docs/default/component/bc-developer-guide/design-system/about-the-design-system/',
-  },
-  {
-    path: '/Data-and-APIs/API-Guidelines',
-    to: '/docs/default/component/bc-developer-guide/bc-government-api-guidelines/',
-  },
-  {
-    path: '/API-Guidelines',
-    to: '/docs/default/component/bc-developer-guide/bc-government-api-guidelines/',
-  },
-];
